fix version sorting
Previously "a0" could be sorted before "a" as we'd try to parse either
the full second string or an empty string (depending on undefined
behavior) as a number, resulting a 0 that's not actually in the string.

This also created undefined behavior as we passed npos to the
remove_prefix function of std::string_view.